HIVE-9337 : Move more hive.spark.* configurations to HiveConf [Spark Branch] (Szehon, reviewed by Brock, Chengxiang, and Lefty)
git-svn-id: https://svn.apache.org/repos/asf/hive/branches/spark@1653432 13f79535-47bb-0310-9956-ffa450edef68

+++ spark-client/src/main/java/org/apache/hive/spark/client/rpc/Rpc.java
@@ -84,7 +84,7 @@ public class Rpc implements Closeable {
       final String secret,
       final RpcDispatcher dispatcher) throws Exception {
     final RpcConfiguration rpcConf = new RpcConfiguration(config);
-    int connectTimeoutMs = rpcConf.getConnectTimeoutMs();
+    int connectTimeoutMs = (int) rpcConf.getConnectTimeoutMs();
 
     final ChannelFuture cf = new Bootstrap()
         .group(eloop)
